Moved around the location of some Event related classes
